+++ src/internals/ir/index.ts
@@ -6,7 +6,9 @@ export {
   CFG,
   getPredecessors,
   getSuccessors,
+  EntryOrigin,
 } from "./cfg";
 export { TactASTStore } from "./astStore";
 export { CompilationUnit, Contract } from "./ir";
 export * from "./types";
+export * from "./builders/tactIRBuilder";